feat(manifest): implement download and processing of files from manifest
